requirements.txt
fastapi>=0.100
uvicorn>=0.23
langchain>=0.3
langchain-groq>=0.2
langchain-community>=0.3
langchain-chroma>=0.1
langgraph>=0.2
langsmith>=0.1
langfuse>=2.0
chromadb>=0.5
python-dotenv>=1.0
sentence-transformers>=3.0
pandas>=2.0
python-multipart>=0.0.6
langchain-huggingface>=0.1
